Support sub-heading in tables
It is a tragedy that docbook has this incredibly elaborate
language for formatting all sorts of things, yet, style sheets
have to resort to processing instructions to do the most basic
table formatting things, like allowing subheadings. Truly
tragic.

So, to mark a table row as subheading (to receive the same
basic formatting as a <thead><row>), add
<?dbhtml class="subhead"?> to the row.

This commit changes style.css to apply the heading style
to such rows.